+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+# v0.3.0
+## 07/05/2017
+
+1. [](#new)
+    * Added Nonce features for printing
+    * Added pf_implode Twig filter
+    * Added configuration option to get breadcrumbs
+    * Added configuration option to get tags
+    * Added configuration option to get first image
+    * Added configuration option to set Close On Escape
+    * Added configuration option to set Close Icon
+    * Added configuration option to set Confirm and Cancel Buttons Text
+    * Added animated icon
+    * Added icons on the buttons
+1. [](#improved)
+    * Improved printfriendly Twig function
+    * Updated jQuery UI library to v1.12.1
+    * Updated Font Awesome library to v4.7.0
+    * English and Italian languages improvements
+    * Deleted superflue javascript library
+1. [](#bugfix)
+    * Fix Broken internal anchors [#2](https://github.com/iusvar/grav-plugin-printfriendly/issues/2)
+
 # v0.2.1
 ## 08/29/2016